# Add support for dotnet SDK fallback folder